# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: NOWFit API Documentation
  version: 1.0.0
  description: >
    Official API documentation for the NOWFit Fitness Coaching Platform.
    Contains all Admin, Coach, Client, Auth, Chat, Diet & Workout APIs.
servers:
  - url: http://localhost:5000
    description: Local Server
...
paths:
...
  /admin/diets/create:
    post:
      tags: [Diets]
      summary: Create a new diet plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
      responses:
        201:
          description: Diet created


  #############################################
  # WORKOUT ROUTES
  #############################################
...